qnd12_app_pro/requirements.txt
django>=3.2,<4.0
django-baton
#Wagtail
wagtail==3.0.3

django-social-share
wagtailgmaps
wagtail-localize
wagalytics==1.3

requests
celery
#gunicorn
uWSGI==2.0.20
redis
django-dotenv
psycopg2-binary
django-storages

#Django Web App dependencies
django-social-auth
social-auth-app-django
sorl-thumbnail
django-widget-tweaks
django-phone-field
django-phonenumber-field[phonenumberslite]
django-bootstrap-datepicker-plus
django-bootstrap-v5
django-bootstrap4
django-jquery
django-embed-video
django-ckeditor
django-qr-code

kombu==5.2.2
django-parler
django-localflavor
django-rosetta
wagtail-localize


boto3
#System HOST
PyOpenSSL
python-dotenv==0.21.0

#Ecommerce
braintree
weasyprint
